Refactor new migration to remove grunt dependency
